+++ src/API_Common.php
@@ -76,7 +76,7 @@ trait API_Common {
 	 * @return string $message License status message.
 	 */
 	public function check_license( $slug ) {
-		$license = trim( get_option( $slug . '_license_key' ) );
+		$license = trim( get_site_option( $slug . '_license_key' ) );
 
 		$api_params = [
 			'edd_action' => 'check_license',
@@ -97,13 +97,13 @@ trait API_Common {
 
 		// We need to update the license status at the same time the message isupdated.
 		if ( $license_data && isset( $license_data->license ) ) {
-			update_option( $slug . '_license_key_status', $license_data->license );
+			update_site_option( $slug . '_license_key_status', $license_data->license );
 		}
 
 		// Get expire date.
 		$expires = false;
 		if ( isset( $license_data->expires ) && 'lifetime' !== $license_data->expires ) {
-			$expires    = date_i18n( get_option( 'date_format' ), strtotime( $license_data->expires, time() ) );
+			$expires    = date_i18n( get_site_option( 'date_format' ), strtotime( $license_data->expires, time() ) );
 			$renew_link = '<a href="' . esc_url( $this->get_renewal_link() ) . '"target="_blank">' . esc_attr( $this->strings['renew'] ) . '</a>';
 		} elseif ( isset( $license_data->expires ) && 'lifetime' === $license_data->expires ) {
 			$expires = 'lifetime';
@@ -152,6 +152,30 @@ trait API_Common {
 	}
 
 	/**
+	 * Constructs a renewal link.
+	 *
+	 * @since 1.0.0
+	 */
+	public function get_renewal_link() {
+		// If a renewal link was passed in the config, use that.
+		if ( ! empty( $this->renew_url ) ) {
+			return $this->renew_url;
+		}
+
+		// If download_id was passed in the config, a renewal link can be constructed.
+		$license_key = trim( get_site_option( $this->slug . '_license_key', false ) );
+		if ( ! empty( $this->download_id ) && $license_key ) {
+			$url  = esc_url( $this->api_url );
+			$url .= '/checkout/?edd_license_key=' . $license_key . '&download_id=' . $this->download_id;
+
+			return $url;
+		}
+
+		// Otherwise return the api_url.
+		return $this->api_url;
+	}
+
+	/**
 	 * Redirect to where we came from.
 	 *
 	 * @param array $error_data Data for error notice.
@@ -178,5 +202,153 @@ trait API_Common {
 
 		wp_safe_redirect( $location );
 		exit();
+	}
+
+	/**
+	 * Disable SSL verification in order to prevent download update failures.
+	 *
+	 * @param  array  $args Array of HTTP args.
+	 * @param  string $url  URL.
+	 * @return object $array
+	 */
+	public function http_request_args( $args, $url ) {
+		if ( false !== strpos( $url, 'https://' ) && strpos( $url, 'edd_action=package_download' ) ) {
+			$args['sslverify'] = $this->verify_ssl();
+		}
+
+		return $args;
+	}
+
+	/**
+	 * Returns if the SSL of the store should be verified.
+	 *
+	 * @since  1.6.13
+	 * @return bool
+	 */
+	protected function verify_ssl() {
+		return (bool) apply_filters( 'edd_sl_api_request_verify_ssl', true, $this );
+	}
+
+	/**
+	 * Calls the API and, if successful, returns the object delivered by the API.
+	 *
+	 * @param  array $data Parameters for the API action.
+	 * @return false|object
+	 */
+	protected function api_request( $data ) {
+		global $edd_plugin_url_available;
+
+		// Do a quick status check on this domain if we haven't already checked it.
+		$store_hash = md5( $this->api_url );
+		if ( ! is_array( $edd_plugin_url_available ) || ! isset( $edd_plugin_url_available[ $store_hash ] ) ) {
+			$test_url_parts = parse_url( $this->api_url );
+
+			$scheme = ! empty( $test_url_parts['scheme'] ) ? $test_url_parts['scheme'] : 'http';
+			$host   = ! empty( $test_url_parts['host'] ) ? $test_url_parts['host'] : '';
+			$port   = ! empty( $test_url_parts['port'] ) ? ':' . $test_url_parts['port'] : '';
+
+			if ( empty( $host ) ) {
+				$edd_plugin_url_available[ $store_hash ] = false;
+			} else {
+				$test_url                                = $scheme . '://' . $host . $port;
+				$response                                = wp_remote_get(
+					$test_url,
+					[
+						'timeout'   => $this->health_check_timeout,
+						'sslverify' => $this->verify_ssl(),
+					]
+				);
+				$edd_plugin_url_available[ $store_hash ] = is_wp_error( $response ) ? false : true;
+			}
+		}
+
+		if ( false === $edd_plugin_url_available[ $store_hash ] ) {
+			return;
+		}
+
+		$data = array_merge( $this->api_data, $data );
+
+		if ( $this->slug !== $data['slug'] ) {
+			return;
+		}
+
+		/**
+		 * Plugins are not able to update from the store.
+		 * This would cause the store to go into maintence mode as the plugin
+		 * tries to update, likely resulting in a non-responsive site.
+		 *
+		 * @link https://github.com/easydigitaldownloads/easy-digital-downloads/issues/7168
+		 */
+		if ( trailingslashit( home_url() ) === $this->api_url ) {
+			return false; // Don't allow a plugin to ping itself.
+		}
+
+		$api_params = [
+			'edd_action' => 'get_version',
+			'license'    => ! empty( $data['license'] ) ? $data['license'] : '',
+			'item_name'  => isset( $data['name'] ) ? $data['name'] : false,
+			'item_id'    => isset( $data['item_id'] ) ? $data['item_id'] : false,
+			'version'    => isset( $data['version'] ) ? $data['version'] : false,
+			'slug'       => $data['slug'],
+			'author'     => $data['author'],
+			'url'        => home_url(),
+			'beta'       => ! empty( $data['beta'] ),
+		];
+
+		$request = $this->get_api_response( $this->api_url, $api_params );
+
+		if ( $request && isset( $request->sections ) ) {
+			$request->sections = maybe_unserialize( $request->sections );
+		} else {
+			$request = false;
+		}
+
+		if ( $request && isset( $request->banners ) ) {
+			$request->banners = maybe_unserialize( $request->banners );
+		}
+
+		if ( $request && isset( $request->icons ) ) {
+			$request->icons = maybe_unserialize( $request->icons );
+		}
+
+		if ( ! empty( $request->sections ) ) {
+			foreach ( $request->sections as $key => $section ) {
+				$request->$key = (array) $section;
+			}
+		}
+
+		return $request;
+	}
+
+	/**
+	 * Disable requests to wp.org repository for this theme.
+	 *
+	 * @since 1.0.0
+	 *
+	 * TODO: Is this necessary, if yes should we implement for plugins too?
+	 *
+	 * @param array  $r   An array of HTTP request arguments.
+	 * @param string $url The request URL.
+	 *
+	 * @return array
+	 */
+	public function disable_wporg_request( $r, $url ) {
+		// If it's not a theme update request, bail.
+		if ( 0 !== strpos( $url, 'https://api.wordpress.org/themes/update-check/1.1/' ) ) {
+			return $r;
+		}
+
+		// Decode the JSON response.
+		$themes = json_decode( $r['body']['themes'] );
+
+		// Remove the active parent and child themes from the check.
+		$parent = get_option( 'template' );
+		$child  = get_option( 'stylesheet' );
+		unset( $themes->themes->$parent, $themes->themes->$child );
+
+		// Encode the updated JSON response.
+		$r['body']['themes'] = wp_json_encode( $themes );
+
+		return $r;
 	}
 }