+++ src/Theme_Updater.php
@@ -14,15 +14,19 @@ namespace EDD\Software_Licensing\Updater;
  */
 class Theme_Updater {
 	use API_Common;
+	use Updater_Common;
 
 	// phpcs:disable Squiz.Commenting.VariableComment.Missing
-	private $api_url      = null;
-	private $response_key = null;
-	private $slug         = null;
-	private $license_key  = null;
-	private $version      = null;
-	private $author       = null;
-	private $strings      = null;
+	private $api_url              = null;
+	private $api_data             = [];
+	private $response_key         = null;
+	private $slug                 = null;
+	private $license_key          = null;
+	private $cache_key            = null;
+	private $version              = null;
+	private $author               = null;
+	private $strings              = null;
+	private $health_check_timeout = 5;
 	// phpcs:enable
 
 	/**
@@ -44,13 +48,15 @@ class Theme_Updater {
 
 		$args = wp_parse_args( $args, $defaults );
 
+		$this->api_url      = $args['api_url'];
+		$this->api_data     = $args;
 		$this->license      = $args['license'];
 		$this->item_name    = $args['item_name'];
 		$this->version      = $args['version'];
 		$this->slug         = sanitize_key( $args['slug'] );
 		$this->author       = $args['author'];
 		$this->beta         = $args['beta'];
-		$this->api_url      = $args['api_url'];
+		$this->cache_key    = $args['cache_key'];
 		$this->response_key = $this->slug . '-' . $this->beta . '-update-response';
 		$this->strings      = $strings;
 	}
@@ -61,135 +67,7 @@ class Theme_Updater {
 	 * @return void
 	 */
 	public function load_hooks() {
-		add_filter( 'site_transient_update_themes', [ $this, 'theme_update_transient' ] );
-		add_filter( 'delete_site_transient_update_themes', [ $this, 'delete_theme_update_transient' ] );
-		add_action( 'load-update-core.php', [ $this, 'delete_theme_update_transient' ] );
-		add_action( 'load-themes.php', [ $this, 'delete_theme_update_transient' ] );
-		add_action( 'load-themes.php', [ $this, 'load_themes_screen' ] );
+		add_filter( 'site_transient_update_themes', [ $this, 'update_transient' ] );
 	}
 
-	/**
-	 * Show the update notification when necessary.
-	 *
-	 * @return void
-	 */
-	public function load_themes_screen() {
-		add_thickbox();
-		add_action( 'admin_notices', [ $this, 'update_nag' ] );
-	}
-
-	/**
-	 * Display the update notifications.
-	 *
-	 * @return void
-	 */
-	public function update_nag() {
-		$theme        = wp_get_theme( $this->slug );
-		$api_response = get_transient( $this->response_key );
-
-		if ( false === $api_response ) {
-			return;
-		}
-
-		$update_url     = wp_nonce_url( 'update.php?action=upgrade-theme&amp;theme=' . rawurlencode( $this->slug ), 'upgrade-theme_' . $this->slug );
-		$update_onclick = ' onclick="if ( confirm(\'' . esc_js( $this->strings['update-notice'] ) . '\') ) {return true;}return false;"';
-
-		if ( version_compare( $this->version, $api_response->new_version, '<' ) ) {
-			echo '<div id="update-nag">';
-			printf(
-				wp_kses_post( $this->strings['update-available'] ),
-				esc_attr( $theme->get( 'Name' ) ),
-				esc_attr( $api_response->new_version ),
-				// $api_response->url . '#TB_inline?width=640&amp;inlineId=' . $this->slug . '_changelog',
-				esc_url( $api_response->url . '&TB_iframe=true&width=1024&width=800' ),
-				esc_attr( $theme->get( 'Name' ) ),
-				esc_url( $update_url ),
-				esc_attr( $update_onclick )
-			);
-			echo '</div>';
-			echo '<div id="' . esc_attr( $this->slug ) . '_changelog" style="display:none;">';
-			echo wp_kses_post( wpautop( $api_response->sections['changelog'] ) );
-			echo '</div>';
-		}
-	}
-
-	/**
-	 * Update the theme update transient with the response from the version check.
-	 *
-	 * @param  array $transient Theme update transient.
-	 * @return array|boolean If an update is available, returns the update parameters.
-	 *                                 If no update is needed returns false.
-	 *                                 If the request fails returns false.
-	 */
-	public function theme_update_transient( $transient ) {
-		$update_data = $this->check_for_update();
-		if ( $update_data ) {
-			// Make sure the theme property is set.
-			// See issue 1463 on Github in the Software Licensing Repo.
-			$update_data['theme'] = $this->slug;
-
-			$transient->response[ $this->slug ] = $update_data;
-		}
-
-		return $transient;
-	}
-
-	/**
-	 * Remove the update data for the theme.
-	 *
-	 * @return void
-	 */
-	public function delete_theme_update_transient() {
-		delete_transient( $this->response_key );
-	}
-
-	/**
-	 * Call the EDD SL API (using the URL in the construct) to get the latest version information.
-	 *
-	 * @return array|boolean If an update is available, returns the update parameters, if no update is needed returns false, if the request fails returns false.
-	 */
-	public function check_for_update() {
-		$update_data = get_transient( $this->response_key );
-
-		if ( false === $update_data ) {
-			$failed = false;
-
-			$api_params = [
-				'edd_action' => 'get_version',
-				'license'    => $this->license,
-				'name'       => $this->item_name,
-				'slug'       => $this->slug,
-				'version'    => $this->version,
-				'author'     => $this->author,
-				'beta'       => $this->beta,
-			];
-
-			$update_data = $this->get_api_response( $this->api_url, $api_params );
-
-			if ( ! is_object( $update_data ) ) {
-				$failed = true;
-			}
-
-			// If the response failed, try again in 30 minutes.
-			if ( $failed ) {
-				$data              = new stdClass();
-				$data->new_version = $this->version;
-				set_transient( $this->response_key, $data, strtotime( '+30 minutes', time() ) );
-
-				return false;
-			}
-
-			// If the status is 'ok', return the update arguments.
-			if ( ! $failed ) {
-				$update_data->sections = maybe_unserialize( $update_data->sections );
-				set_transient( $this->response_key, $update_data, strtotime( '+12 hours', time() ) );
-			}
-		}
-
-		if ( version_compare( $this->version, $update_data->new_version, '>=' ) ) {
-			return false;
-		}
-
-		return (array) $update_data;
-	}
 }